requirements.txt
alpaca-trade-api
requests
python-dotenv
pytz
yfinance
numpy
beautifulsoup4
matplotlib
pandas